requirements.txt
streamlit
pymongo
python-dotenv
pandas
scikit-learn
transformers
torch
reportlab
pymupdf
Pillow
pytesseract
langchain
langchain-google-genai
google-generativeai